Optimize TM::String::prepend
The old code created a temporary buffer on the stack, copied the data
from m_str to the buffer, and copied the buffer back to m_str (which
might have been relocated at that point). The change gets rid of the
temporary buffer, which removes one of the memcpy operations.

As a fortunate side effect: this fixes a bug in case on of the
TM::String objects had a nullbyte embedded. The old code did use
`set_str(char*)`, which uses `strlen` to determine the length of the
argument.

`prepend("\0")` will still fail, you would either have to wrap the input
char* in a TM::String object with a length, or add an overload for
`TM::String::prepend(char*, size_t)` to pass the length as an argument.