Reduce duplication in TM::String::set_str
There are two variants of this method: one with a length argument, and
one without a length argument. The second one uses `strlen` to calculate
the length. Forward this call to the variant with a length argument and
reduce the code duplication.
The compiler will most likely inline this call, which means there is no
performance penalty.