Implement operator+(=) for String and StringView
We can override these as free functions, so now we have a nice way to
concatenate a StringView to a String without the need to reach to raw
pointers or converting the StringView to a temporary String first.

A simple script comparing `str1.append(sv.to_string());` to
`str1 += sv.to_string();` has a bit over 400 fewer instructions. It also
removes an unessecary memory allocation.